// web/miniapp/src/shared/api/house.ts
export const NotificationPreferencesSchema = z.object({
 notifications_enabled: z.boolean(),
 issue_notifications_enabled: z.boolean(),
 announcement_notifications_enabled: z.boolean(),
 membership_notifications_enabled: z.boolean(),
 bot_notifications_enabled: z.boolean(),
});
export type NotificationPreferences = z.infer<typeof NotificationPreferencesSchema>;